third_party/WebKit/LayoutTests/bluetooth/characteristic/startNotifications/gen-gatt-op-device-reconnects-during-error.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
promise_test(() => {
  let val = new Uint8Array([1]);
  return setBluetoothFakeAdapter('DisconnectingDuringFailureGATTOperationAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['health_thermometer']}]}))
    .then(device => device.gatt.connect())
    .then(gatt => gatt.getPrimaryService('health_thermometer'))
    .then(service => service.getCharacteristic('measurement_interval'))
    .then(characteristic => {
      let disconnected = eventPromise(characteristic.service.device, 'gattserverdisconnected');
      let promise = assert_promise_rejects_with_message(
        characteristic.startNotifications(),
        new DOMException('GATT Server disconnected while performing a GATT operation.',
                         'NetworkError'));
      return disconnected.then(() => characteristic.service.device.gatt.connect())
        .then(() => promise);
    });
}, 'Device reconnects during a startNotifications call that fails. Reject ' +
   'with NetworkError.');

</script>
